Moved FieldDefinition out of BulkGenerator
